# 8283996: Reduce cost of year and month calculations